clean up common
