+++ solution/backend/content_search/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from common.fields import HeadlineField
-from common.serializers import DetailsSerializer
+from common.serializers.mix import DetailsSerializer
 from file_manager.serializers.groupings import DocumentTypeSerializer, SubjectSerializer
