+++ solution/backend/resources/serializers/locations.py
@@ -1,7 +1,7 @@
 from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
 
-from common.mixins import PolymorphicSerializer, PolymorphicTypeField
+from common.serializers.clean import PolymorphicSerializer, PolymorphicTypeField
 from common.utils import ProxySerializerWrapper
 from resources.models import Section, Subpart
